requirements.txt
torch
torchvision
timm
opencv-python-headless
pillow
numpy
gdown
gradio
